fix: outlier u64 usage

+++ rln/src/protocol.rs
@@ -156,18 +156,18 @@ pub fn proof_inputs_to_rln_witness(
     let (identity_secret, read) = bytes_le_to_fr(&serialized[all_read..]);
     all_read += read;
 
-    let id_index = u64::from_le_bytes(serialized[all_read..all_read + 8].try_into()?);
+    let id_index = usize::from_le_bytes(serialized[all_read..all_read + 8].try_into()?);
     all_read += 8;
 
     let (epoch, read) = bytes_le_to_fr(&serialized[all_read..]);
     all_read += read;
 
-    let signal_len = u64::from_le_bytes(serialized[all_read..all_read + 8].try_into()?);
+    let signal_len = usize::from_le_bytes(serialized[all_read..all_read + 8].try_into()?);
     all_read += 8;
 
-    let signal: Vec<u8> = serialized[all_read..all_read + (signal_len as usize)].to_vec();
+    let signal: Vec<u8> = serialized[all_read..all_read + signal_len].to_vec();
 
-    let merkle_proof = tree.proof(id_index as usize).expect("proof should exist");
+    let merkle_proof = tree.proof(id_index).expect("proof should exist");
     let path_elements = merkle_proof.get_path_elements();
     let identity_path_index = merkle_proof.get_path_index();
